Move the implementation from api.rest to services
The services package is a better description of the classes than the
old api.rest.

KeyProviderService and Service84JWTAFConfigurer package name are not
part of the external API so we can change the package without a major
version increment.

Service84JWTAuthenticationFilter in the api.rest package is part of the
external api so we will move the implementation to services and make use
inheritance to maintain the external API.